de-componentized tabs on details
